requirements.txt
anthropic>=0.34.0
python-dotenv>=1.0.0
numpy>=1.24.0
scipy>=1.10.0
tqdm>=4.65.0
ipykernel
matplotlib>=3.7.0
seaborn>=0.12.0
httpx>=0.27.0
pandas>=2.0.0